cxc/requirements.txt
langchain_core==1.2.5
langchain_openai==1.1.6
matplotlib==3.10.8
Pillow==12.0.0
torch==2.9.1+cu128
tqdm==4.67.1
transformers==4.57.3
